Added MAX_GROUP_LOADS

+++ services/src/main/java/fr/ippon/tatami/service/TimelineService.java
@@ -571,8 +571,7 @@ public class TimelineService {
     private Collection<StatusDTO> getUserline(String login, int nbStatus, String start, String finish, int rounds) {
         List<String> statuses = userlineRepository.getUserline(login, nbStatus, start, finish);
         Collection<StatusDTO> dtos = buildStatusList(statuses);
-        if (rounds < Constants.MAX_GROUP_LOADS)
-        if (statuses.size() != dtos.size()) {
+        if (rounds < Constants.MAX_GROUP_LOADS && statuses.size() != dtos.size()) {
             Collection<String> statusIdsToDelete = findStatusesToCleanUp(statuses, dtos);
             userlineRepository.removeStatusesFromUserline(login, statusIdsToDelete);
             if (statusIdsToDelete != null && !statusIdsToDelete.isEmpty()) {
